Library

Module

AuditoryStimuli.AuditoryStimuli — Module

A Julia package for generating auditory stimuli.

Signal generators

AuditoryStimuli.NoiseSource — Type

NoiseSource(eltype, samplerate, nchannels, std=1)

NoiseSource is a multi-channel noise signal generator. The noise on each channel is independent.

Inputs

samplerate specifies the sample rate of the signal specified in Hz.

nchannels specifies the number of channels of the signal.

std specifies the desired standard deviation of the signal.

Output

SampleSource object

Example

source_object = NoiseSource(Float64, 48u"kHz", 2, 0.3)
wn = read(source_object, 480)         # Specify number of samples of signal to generate
wn = read(source_object, 50u"ms")     # Specify length of time of signal to generate

AuditoryStimuli.CorrelatedNoiseSource — Type

CorrelatedNoiseSource(eltype, samplerate, nchannels, std, correlation)

CorrelatedNoiseSource is a multi-channel (currently restricted to 2) noise signal generator.

Inputs

samplerate specifies the sample rate of the signal.

nchannels specifies the number of channels of the signal.

std specifies the desired standard deviation of the signal.

correlation specifies the desired correlation between the signals.

Output

SampleSource object

Example

source_object = CorrelatedNoiseSource(Float64, 48000, 2, 0.3, 0.75)
cn = read(source_object, 480)         # Specify number of samples of signal to generate
cn = read(source_object, 50u"ms")     # Specify length of time of signal to generate

Issues

Currently only supports 2 channels

AuditoryStimuli.HarmonicComplex — Type

HarmonicComplex(eltype, samplerate, freqs)

HarmonicComplex is a single-channel sine-tone signal generator. freqs can be an array of frequencies for a multi-frequency source, or a single frequency for a mono source.

Online signal modifiers

AuditoryStimuli.Amplification — Type

Amplification(target, current, change_limit)

Apply amplification to the signal

Inputs

target desired linear amplification factor to be applied to signal.

current linear amplification currently applied to signal. Also used to specify the intial value for the process.

amplification_step_change_limit maximum change that can occur per frame.

Output

SampleBuf

Example

amplify = Amplification(0.1, 0.0, 0.05)
attenuated_sound = write(amplify, original_sound)

AuditoryStimuli.Filter — Type

Filter(filters)

Apply filter to the signal

Inputs

filters array of DSP filter objects.

Output

Filter object

Example

using DSP
responsetype = Bandpass(500, 4000; fs=48000)
designmethod = Butterworth(4)
zpg = digitalfilter(responsetype, designmethod)
f_left = DSP.Filters.DF2TFilter(zpg)
f_right = DSP.Filters.DF2TFilter(zpg)
bandpass = AuditoryStimuli.Filter([f_left, f_right])

Signal modifiers

AuditoryStimuli.bandpass_noise — Function

bandpass_noise(number_samples, number_channels, lower_bound, upper_bound, sample_rate; filter_order=14)

Generates band pass noise with specified upper and lower bounds using a butterworth filter.

AuditoryStimuli.bandpass_filter — Function

bandpass_filter(AbstractArray, lower_bound, upper_bound, sample_rate; filter_order=14)
bandpass_filter(SampledSignal, lower_bound, upper_bound;              filter_order=14)

Signal will be filtered with bandpass butterworth filter between 'lowerbound' and `upperboundwith filter offilter_order`.

AuditoryStimuli.ramp_on — Function

ramp_on(data, number_samples)

Apply a linear ramp to start of signal

AuditoryStimuli.ramp_off — Function

ramp_off(data, number_samples)

Apply a linear ramp to end of signal

AuditoryStimuli.set_RMS — Function

set_RMS(data, desired_rms)

Modify rms of signal to desired value

AuditoryStimuli.set_ITD — Function

set_ITD(data, number_samples)

Introduce an ITD of number_samples

AuditoryStimuli.amplitude_modulate — Function

amplitude_modulate(data, modulation_frequency, sample_rate; phase=π)

Amplitude modulates the signal

See wikipedia

AuditoryStimuli.ITD_modulate — Function

ITD_modulate(data, modulation_frequency, ITD_1, ITD_2, samplerate)

Modulate an applied ITD

Plotting

AuditoryStimuli.PlotSpectroTemporal — Function

PlotSpectroTemporal(data, sample_rate)

This function plots the time, spectrogram, and periodogram of a signal